% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixedincome.R
\name{bootstrap}
\alias{bootstrap}
\title{Bootstrapping}
\usage{
bootstrap(data, samp, size, remp = TRUE)
}
\arguments{
\item{data}{is the sample population.}

\item{samp}{is the number of samples to create.}

\item{size}{is the size of the generated samples.}

\item{remp}{sampling with replacement, default = TRUE.}
}
\value{
a matrix with each sample as a column of the matrix.
}
\description{
Calculates different samples from the sample population.
}
