% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posiciones.R
\name{diah}
\alias{diah}
\title{DiaH}
\usage{
diah(fecha, tipo = "Habil")
}
\arguments{
\item{fecha}{is a date in %Y-%m-%d format.}

\item{tipo}{"Last" returns the last business day, any other value returns if the given date is business date.}
}
\value{
a dataframe with dates and prices as requested.
}
\description{
Function that tells if the given date is business date or returns the last business date.
}
