% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixedincome.R
\name{Duration}
\alias{Duration}
\title{Duration of a Bond}
\usage{
Duration(mat, day, tcoupn, ytm)
}
\arguments{
\item{mat}{is the maturity date of the Bond}

\item{day}{is the day in which the price is calculated}

\item{tcoupn}{is the coupon rate of the bond (annualized)}

\item{ytm}{is the Yield to Maturity of the Bond}
}
\value{
the Duration of the Bond
}
\description{
Calculates the Duration of a Bond with the maturity date, calculation date, coupon rate and YTM of the Bond
}
