% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflation.R
\name{fechas_inflacion}
\alias{fechas_inflacion}
\title{Dates for inflation}
\usage{
fechas_inflacion(fecha)
}
\arguments{
\item{fecha}{is the date for which the dates will be calculated, format = YYYY-mm-dd.}
}
\value{
the inflation rate for the given dates and type.
}
\description{
Calculates the date a month ago, a quarter ago, a semester ago and a yera ago.
}
