% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prices_dates.R
\name{get_bonds}
\alias{get_bonds}
\title{Get Bonds}
\usage{
get_bonds(id = NULL)
}
\arguments{
\item{id}{is the id vector of which bonds information must be obtained.}
}
\value{
a dataframe with Issuing date, Maturity date, Coupon rate, Type of rate, Spread and Grade of the selected Bonds.
}
\description{
Function that does a query from a mysql database in order to obtain characteristics of bonds.
}
