% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prices_dates.R
\name{get_indices}
\alias{get_indices}
\title{Get Indices}
\usage{
get_indices(fecha = NULL, id = NULL)
}
\arguments{
\item{fecha}{is the dates vector of which the nodes are needed.}

\item{id}{is the id vector of which the level information must be obtained.}
}
\value{
a dataframe with nodes id, date and value.
}
\description{
Function that does a query from a mysql database in order to obtain a level of an index.
}
