% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prices_dates.R
\name{get_prices}
\alias{get_prices}
\title{Get Prices}
\usage{
get_prices(fecha = NULL, id = NULL)
}
\arguments{
\item{fecha}{is a date vector which can be build with seq_Date() function.}

\item{id}{is the id vector of which the prices must be obtained.}
}
\value{
a dataframe with dates and prices as requested.
}
\description{
Function that does a query from a mysql database in order to obtain prices of instruments.
}
