% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixedincome.R
\name{Grade2Number}
\alias{Grade2Number}
\title{Grade2Number}
\usage{
Grade2Number(note)
}
\arguments{
\item{note}{is the grade assigned by the rating agencies.}
}
\value{
a comparable number for each note.
}
\description{
Calculates a comparable number for Fitch, Moody's, S&P and HR Ratings grades.
}
