% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{greet}
\alias{greet}
\title{Greet someone}
\usage{
greet(name = "")
}
\arguments{
\item{name}{a string with the name of whom to greet}
}
\value{
a string with the greeting
}
\description{
Receives a name a greets the person
}
