% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflation.R
\name{Inflacion1}
\alias{Inflacion1}
\title{Inflation rate over a given period of time}
\usage{
Inflacion1(inicio, fin, tipo = "general")
}
\arguments{
\item{inicio}{starting period for inflation calculation.}

\item{fin}{ending period for inflation calculation,}

\item{tipo}{it can be: general, subyacente, nosubyacente.}
}
\value{
the inflation rate for the given dates and type.
}
\description{
Calculates the inflation rate for the selected date.
}
