% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflation.R
\name{Inflacion2}
\alias{Inflacion2}
\title{Inflation rate within a given window for selected dates.}
\usage{
Inflacion2(fechas, tipo = "general")
}
\arguments{
\item{fechas}{is a vector with the dates on which inflation will be calculated, they must be in "YYYY-mm" format.}

\item{tipo}{it can be: general, subyacente, nosubyacente.}
}
\value{
the inflation rate for the given dates and type.
}
\description{
Calculates the Monthly, Quarterly, Semianually and Anually inflation rate.
}
