% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixedincome.R
\name{ModifiedDuration}
\alias{ModifiedDuration}
\title{Modified Duration of a Bond}
\usage{
ModifiedDuration(mat, day, tcoupn, sobretasa, yield, period = 182)
}
\arguments{
\item{mat}{is the maturity date of the Bond}

\item{day}{is the day in which the price is calculated}

\item{tcoupn}{is the coupon rate of the bond (annualized)}

\item{sobretasa}{is the spread of the coupon.}

\item{yield}{is the Yield to Maturity of the Bond}

\item{period}{is the time for every coupon payment}
}
\value{
the  Modified Duration of the Bond
}
\description{
Calculates the Duration of a Bond with the maturity date, calculation date, coupon rate and YTM of the Bond
}
