% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixedincome.R
\name{Number2Grade}
\alias{Number2Grade}
\title{Number2Grade}
\usage{
Number2Grade(number)
}
\arguments{
\item{number}{is the number given by the user.}
}
\value{
a grade comparable with Fitch, Moody's, S&P and HR Ratings grades.
}
\description{
Calculates a grade for a given number
}
