% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixedincome.R
\name{PortfolioDuration}
\alias{PortfolioDuration}
\title{Modified Duration of a Bond Portfolio}
\usage{
PortfolioDuration(date = Sys.Date() - 1, instruments, weight)
}
\arguments{
\item{date}{is the date in which the Portfolio Duration is going to be calculated.}

\item{instruments}{is an array containing the instrument's id.}

\item{weight}{is an array containing the instrument's weights in the portfolio.}
}
\value{
Modified Duration of the Portfolio.
}
\description{
Calculates the Duration of a Bond portfolio with the id's of the instruments and it's weights on the portfolio.
}
