% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixedincome.R
\name{PriceChange}
\alias{PriceChange}
\title{Price Change of a Bond}
\usage{
PriceChange(mat, day, tcoupn, sobretasa, yield, period = 182,
  cyield = 1e-04)
}
\arguments{
\item{mat}{is the maturity date of the Bond}

\item{day}{is the day in which the price is calculated}

\item{tcoupn}{is the coupon rate of the bond (annualized)}

\item{sobretasa}{is the spread of the coupon.}

\item{yield}{is the Yield to Maturity of the Bond}

\item{period}{is the time for every coupon payment}

\item{cyield}{is the change in the yield, by default is set to 1 bps.}
}
\value{
the Price Change in the Bond
}
\description{
Calculates the Price Change of a Bond with the maturity date, calculation date, coupon rate, YTM of the Bond and Yield change
}
