% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prices_dates.R
\name{seq_Date}
\alias{seq_Date}
\title{Sequence of Dates}
\usage{
seq_Date(fecha = NULL)
}
\arguments{
\item{fecha}{is a character: "20150101/" makes a sequence of dates begining at 2015-01-01, "20150101/20160101" makes a sequence of dates begining at 2015-01-01 ending at 2016-01-01}
}
\value{
a dataframe with dates and prices as requested.
}
\description{
Function that obtains a sequence of dates.
}
