% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskreturn.R
\name{ValueAtRisk}
\alias{ValueAtRisk}
\title{Historical VaR of a Security's Portfolio}
\usage{
ValueAtRisk(instruments, shares, confidence = 0.95, period = 252)
}
\arguments{
\item{instruments}{is an array of instruments.}

\item{shares}{is an array of each instrument's shares.}

\item{confidence}{is the level of confidence at which the VaR is calculated.}

\item{period}{es the period in days for the VaR calculation.}
}
\value{
VaR of the security's portfolio
}
\description{
Calculates the VaR metric of a security's portfolio
}
