% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prices_dates.R
\name{get_rates}
\alias{get_rates}
\title{Get the node of the dates needed for a specified rate.}
\usage{
get_rates(fecha = NULL, id = NULL)
}
\arguments{
\item{fecha}{is the dates vector of which the nodes are needed.}

\item{id}{is the id vector of which nodes information must be obtained.}
}
\value{
a dataframe with nodes id, date and value.
}
\description{
Function that does a query from a mysql database in order to obtain the node of some rates
}
