% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskreturn.R
\name{RiskValues}
\alias{RiskValues}
\title{Historical VaR and CVaR of a Security's Portfolio}
\usage{
RiskValues(fecha = Sys.Date() - 1, instruments, shares, efectivo,
  confidence = 0.95, period = 252)
}
\arguments{
\item{fecha}{is the date of calculation.}

\item{instruments}{is an array of instruments.}

\item{shares}{is an array of each instrument's shares.}

\item{efectivo}{is the quantity of cash in the portfolio.}

\item{confidence}{is the level of confidence at which the VaR and CVaR are calculated.}

\item{period}{es the period in days for the VaR and CVaR calculation.}
}
\value{
VaR and CVaR of the security's portfolio
}
\description{
Calculates the VaR metric of a security's portfolio
}
