% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixedincome.R
\name{YTM}
\alias{YTM}
\title{Yield to Maturity of a Bond}
\usage{
YTM(mat, day, tcoupn, precio, period = 182)
}
\arguments{
\item{mat}{is the maturity date of the Bond}

\item{day}{is the day in which the price is calculated}

\item{tcoupn}{is the coupon rate of the bond (annualized)}

\item{precio}{is the price of the bond}

\item{period}{is the time for every coupon payment.}
}
\value{
the YTM of the Bond at the calculation date
}
\description{
Calculates the YTM of a Bond with the maturity date, calculation date, coupon rate and price of the Bond
}
